+++ packages/cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Implement `hi18n sync --check` (shorthand: `-c`) option to raise an error when files would be changed.
+  It is useful when you want to ensure synchronization in your CI.
+
 ## 0.1.4
 
 - Placeholder is changed from `msg()` to `msg.todo("[TODO: example/greeting]")`.